+++ addOns/openapi/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The "Import an OpenAPI definition from the local file system" and "Import an OpenAPI definition from a URL" menu items
   were merged into one, "Import an OpenAPI Definition".
 - Depend on newer versions of Automation Framework and Common Library add-ons (Related to Issue 7961).
+- Use Common Library add-on to obtain the Value Generator (Issue 8016).
 
 ### Fixed
 - Importing empty or invalid OpenAPI definitions failed silently in some cases (Issue 7949).